Improves comments; cleans up code.
